+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_s3_policies.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2022-2025, NetApp, Inc
+# (c) 2022-2026, NetApp, Inc
 # GNU General Public License v3.0+
 # (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
@@ -82,6 +82,31 @@ options:
           - allow
           - deny
         required: true
+
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.4.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality. See README for example usage.
 '''
 EXAMPLES = """
 - name: Create and modify a S3 policy
@@ -153,9 +178,13 @@ class NetAppOntapS3Policies:
                 actions=dict(required=True, type='list', elements='str'),
                 effect=dict(required=True, type='str', choices=['allow', 'deny']),
             ))))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
 
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
+            required_if=[
+                ['use_lambda', True, ('lambda_config',)]
+            ],
             supports_check_mode=True
         )
         self.svm_uuid = None
@@ -180,7 +209,7 @@ class NetAppOntapS3Policies:
                                   exception=traceback.format_exc())
         # sid is an Str or a number, it will return a string back unless you pass a number then it returns a int
         if record:
-            for each in record['statements']:
+            for each in record.get('statements', []):
                 each['sid'] = str(each['sid'])
         return record